PLT-193 Workflow for github-actions calling reusable workflow (#16)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-193

## 🛠 Changes

Added a workflow for github-actions terraform that calls the reusable
terraform-fmt workflow.

## ℹ️ Context for reviewers

This is a minimal proof for testing out reusable workflows, and allows
for exploring some ideas to expand in #13.

## ✅ Acceptance Validation

Testing workflows before the PR is merged is tricky. We'll likely need
to test after merge, then create a follow-up PR if there are issues.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.