DPC-3767 Add infrastructure for DPC WAF-sync lambda (#127)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-3767

## 🛠 Changes

- Added wav-sync as service
- Added plan and apply wav-sync github actions

## ℹ️ Context

DPC needs to synchronize ip addresses in a database with an AWS WAF IP
Set. We [decided](https://confluence.cms.gov/x/1ZUZQQ) to use a
cron-triggered lambda to manage this. This ticket creates the
infrastructure for automated deployment of the lambda.

The aws_waf_access inline policy seems to be the minimum necessary for
the lambda to run. It may require updating (the lambda has not been
written).

The name of the IP Set was intuited by browsing the WAF IP Sets in AWS
Console.

The terraform scripts and github actions are modified versions of the
opt-out-import service.

## 🧪 Validation

- Comparison of the plan output to opt-out-import plan output show that
the permissions necessary are present.
- Plan applied successfully