PLT-503: Add test and sandbox environments to DPC WAF plan and apply (#128)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-503

## 🛠 Changes

Adds test and sbx environments to the DPC WAF configuration

## ℹ️ Context

These are changes made as a part of the overall WAF migration work.

## 🧪 Validation

Once this is applied, we should see the Web ACL configurations show up
with the placeholder IP sets. We'll need to manually update them in AWS,
and then reapply once we remove the association with the security group
on ingress.