PLT-504: Add BCDA dev environment to WAF configuration (#129)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-504

## 🛠 Changes

This adds the BCDA dev environment to the apply workflow of the new WAF
configuration.

## ℹ️ Context

We're moving all apps to WAF for IP filtering. This should alleviate
some of the issues we have with security groups and maintaining our CIDR
whitelists.

## 🧪 Validation

BCDA dev should be able to be reached on the VPN, but no other traffic
should be let through. The BCDA team may need to add CIDR ranges to the
new IP-set that will be created. We'll also need to remove the
association to the dev WAF and security group in place in the bcda-ops
repo: https://github.com/CMSgov/bcda-ops/pull/1114