PLT-724 Apply api-rds service to ab2d dev (#140)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-724

## 🛠 Changes

Added dev environment to plan/apply action matrices

## ℹ️ Context

PLT-719 needs for us to update ab2d dev and test to Postgres 16, this is
attempting to make that change from the same location. for both .

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Validated locally with terraform plan; also verifying via terraform plan
github action