PLT-724 - Step 2.2 creating parameter group for v16 upgrade (#141)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-724

## 🛠 Changes

Relevant to PLT-728 as well, this ticket completes step 2.2 of the work
plan, creating the v16 parameter group to be consumed by the "green"
deployment of ab2d dev and test

## ℹ️ Context

Following work plan, the new environments will need a separate parameter
group to handle the v16 version of their parameters. While there are no
breaking changes, this allows us to separate the parameters of v15
instances and v16 instances.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local terraform plan, validated with GHA plan