PLT-754 Ensure runner runs as root (#166)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-754

## 🛠 Changes

Ensure the runner container runs as root.

## ℹ️ Context

As noted in https://github.com/actions/checkout/issues/956, GitHub still
expects runner containers to run as root. Permissions around the
actions/checkout step break in a number of ways if this is not set.

## 🧪 Validation

Will apply and test in the job currently failing on
https://github.com/CMSgov/bcda-app/pull/1031.