PLT-948: Add clone to temp directory (#189)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-948

## 🛠 Changes

Changes the git clone method to one where we clone to a temp directory
to get around git clone's limitation of needing the directory to be
empty

## ℹ️ Context

After implementing a fix where we check to see if tfenv is already
installed, we were still getting errors with `git clone` because the
directory already existed.

## 🧪 Validation

Workflows should pass and not be stopped by previous tfenv
installations.