PLT-785: Add AB2D Sandbox to WAF environments (#170)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-785

## 🛠 Changes

Adds AB2D sandbox to the list of environments we deploy the shared WAF
terraform to.

## ℹ️ Context

We want to test the changes in all environments before we rollout to
production. The new shared terraform will bring all the teams to one
consistent state

## 🧪 Validation

AB2D sandbox access should remain as usual.