PLT-564: Apply shared terraform for database to BCDA dev, sbx, prod (#192)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-564

## 🛠 Changes

modification made to code for bcda opensbx env and env added to plan and
apply workflow

## ℹ️ Context

need to import bcda-sbx rds resources to platform repo terraform state 

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

output in GitHub actions for the opensbx env

---------

Co-authored-by: Maboh Christopher <chris.maboh@aquia.io>