PLT-709: Add AB2D Prod to Apply Matrix (#197)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-709

## 🛠 Changes

Adds ab2d to the apply workflow matrix.

## ℹ️ Context

As part of the rollout of the WAF shared service to AB2D prod, we have
to add the environment to the terraform apply for the service.

## 🧪 Validation

Once this and https://github.com/CMSgov/ab2d-ops/pull/313 are merged,
the new WAF shared service will be associated with the existing load
balancer, and handle all IP filtering for the AB2D API.