PLT-508: Add Prod environment to matrix (#193)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-508

## 🛠 Changes

Adds the prod environment to the action's environment matrix.

## ℹ️ Context

When the initial PR for PLT-508 was merged, it only included dev and
test for testing purposes, once those pass, we should be able to merge
this PR in, but only after the WAF shared service has been rolled out to
AB2D production.

## 🧪 Validation

The script should update the WAF IP sets for each environment with the
current state of the allowed_ips list in the private repo.