Remove openssl workaround (#200)
This is to remove the workaround we needed for openssl 3.0.8

bugfix


## 🛠 Changes

This is to remove the workaround we implemented for openssl 3.0.8

## ℹ️ Context

Amazon released the latest openssl 3.2.2 and it looks like we are
hitting errors potentially due to this workaround.

## 🧪 Validation

We will merge and test revert if it does not work.