PLT-506: External services WAF IP sync (#196)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-506

## 🛠 Changes

- Added new file external-services-ip-sets-sync.yml runner file
- Added new file waf-ip-set-sync-external-services.sh to run the sync
via the AWS CLI.
- Added new file allowed-ips.txt under
cdap-private/ip-sets/external-services with IP sets taken from BCDA
account.

## ℹ️ Context

These changes are intended to establish a framework for syncing the
external-services IPSets from various accounts and environments to a
file of CIDR IP ranges in the CDAP private repo. This workflow is
currently configured to be manually run, but can easily be reconfigured
to run on a schedule or on an update to the CDAP private repo.

## 🧪 Validation

These changes were validated by running a GHA workflow against the AB2D
IMPL and BCDA/DPC accounts us-east-1 regions.