[PLT-890] add legacy variable to tfstate service and create backends (#212)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-890

## 🛠 Changes

backend files defining s3 state buckets for greenfield apps

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

Add backends to the TF in the platform repo (each backend is a 2-line
file for each of the vpcs). We expect to have 12 new VPCs.

## 🧪 Validation

Buckets are created in S3 with tfstate files for each new backend
definition file.

---------

Co-authored-by: Grant Freeman <129095098+gfreeman-navapbc@users.noreply.github.com>
Co-authored-by: Sean Fern <seanfern@navapbc.com>