PLT-977 Create parameter store and secrets manager migration scripts (#211)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-977

## 🛠 Changes

- created a new 'tools' directory in the platform repo.
- added migrate_parameters and migrate_secrets Python scripts.
- added a settings.ini file to contain required configuration info.
- added a README with basic usage instructions.

## ℹ️ Context

This PR is intended to preserve the scripts that were developed for the
purpose of migrating parameters and secrets from the legacy BCDA and
AB2D AWS accounts to the Greenfield prod and non-prod accounts.

While this PR was opened against the Platform repo, it is not clear if
this is the desired repo for an internal tools directory. If a repo such
as cdap-private is more appropriate for tools then this PR can be
recreated in that repo.

## 🧪 Validation

These scripts were tested with a series of dry-runs followed by some
test migrations of very limited subsets of parameters and secrets (i.e.
> 3).

The full migration has now been run, and further validation in the form
of spot-checking various parameters and secrets was conducted.