PLT-1040: Add bucket SSM Parameter back to Greenfield Environments (#223)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1040

## 🛠 Changes

This adds back an optional parameter that references the name of the
bucket passed in through the `function` module.

## ℹ️ Context

We removed the SSM Parameter for the bucket since it was creating
parameters in every environment on legacy as well as greenfield. This
removal makes greenfield development a bit harder as we cannot
interpolate the bucket name by by App and Environment during deployment.

## 🧪 Validation

This should allow DPC to continue greenfield development and deployment