PLT-1008: Support API RDS in Greenfield (#224)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1008

## 🛠 Changes

In short, this PR brings about necessary changes to run in the
greenfield environments along with some bug fixes that weren't revealed
until now:
- fix the AWS secret lookup resolution; document
- make use of default aws/rds KMS keys
- fix issues where bcda terraform and _this_ module both attempt to
create local hosted zones
- introduce a number of `rds-api` module dependency security groups to
`service-security-groups` for greenfield
- make required variables in `rds-api` optional using lookups (mgmt
cidr) or ignoring them (jenkins subnets)
- annotate and disable certain resources in greenfield to allow this
module to apply


## ℹ️ Context
This is the first time we've had the opportunity to use these modules to
instantiate environments.
Naturally, there were a number of key assumptions that made it into the
solution.
Where possible, we've gone forward with temporarily disabling certain
resource dependencies that have not yet been met in the greenfield
environment fully understanding that some of these dependencies will
never be met.

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

CI should run with results that have few, modest changes to the
resources. Namely, we're looking for tags on legacy resources to be
updated, and little else.