PLT-1033: Introduce a platform child module (#231)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1033

## 🛠 Changes

Introduces a child module for commonly referenced security groups, iam
roles, subnets, and other basics that any consuming module might
otherwise lookup for itself with a separate collection of data sources
and input parameters. 😬

## ℹ️ Context

This will help define the standards by which the CDAP customers operate,
including setting a baseline for default tags.

## 🧪 Validation

Because this is for greenfield only, I've been able to make use of this
child module in a limited capacity in the AB2D core service module–it
works!