PLT-1033: Introduce, harmonize standards module (#233)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1033

## 🛠 Changes
- introduce internal-use only (CDAP) standards module
- address minor issues introduced as part of PLT-1060
- document limitations of the platform module
- initial use of standards/platform modules in `api-rds`

## ℹ️ Context
Per commentary on #231 (the PR that introduced the `platform` module),
it was clear that CDAP needed a secondary module for internal use only
to at least set standard tagging definitions. This PR addresses those
comments.

## 🧪 Validation
- Cleanly applies in legacy environments with expected updates to
standard tagging
- There are clean applies for existing AB2D greenfield environments,
however #234 will is a fast follow to ensure that dpc greenfield applies
correctly in greenfield.