PLT-1040 Extend tf-api-rds workflow to other environments (#240)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1040

## 🛠 Changes

Extend the tf-api-rds workflow to plan and apply in all greenfield
environments.

## ℹ️ Context

This follows up on #234 to include environments that are not yet active.
Once teams have restored databases and added secrets in successive
environments, this workflow can be run again to apply terraform.

## 🧪 Validation

See checks. Plan and apply on environments that have not yet been built
out for rds are expected to fail.