PLT-1130:Create Roles and Policies for each API (#245)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1130

## 🛠 Changes

Snyk role and policy was modified to 3 separate roles and policies per
team

## ℹ️ Context

Given all three APIs are in the same account now (looking at BCDA org,
we have AB2D and DPC ECR repositories which are different org) It will
be nice to have some isolation for API teams

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

See checks.

---------

Co-authored-by: Maboh Christopher <chris.maboh@aquia.io>