PLT-1131: Add files for NAT gateways to external-services sync workflow (#247)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1131

## 🛠 Changes

Adds two files, one for each account, to the external-services sync
workflow.

## ℹ️ Context

When applied, this will add the public NAT gateway IPs to the
external-services allowlist to let the teams reach their public load
balancers from resources on the same account.

## 🧪 Validation

Once this is applied, we should see the IPs added in
https://github.com/CMSgov/cdap-private/pull/8 in the external-services
IP set.