BCDA-9143: Update aco creds bucket name for GF (#252)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9143

## 🛠 Changes

The aco creds output bucket name is changed for Greenfield.
The IAM policies for the lambda are updated to remove encryption/KMS
rules for Greenfield.

## ℹ️ Context

Our bucket naming strategy and s3 encryption management have changed for
greenfield, and the Create ACO Creds lambda must be updated accordingly.
The legacy settings were left untouched, and they can be removed when
the rest of the legacy variables are removed.

In particular, in our legacy infrastructure we managed our s3 encryption
ourselves, including a key and roles that were assigned to the lambda
function. In greenfield, this encryption is handled by default by AWS,
so we can remove a lot of our custom configuration.

Related PR which added the bucket to greenfield:
https://github.com/CMSgov/bcda-ops/pull/1220

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->