DPC-4713: SNS To Slack lambda (#259)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4713

## 🛠 Changes

Added lambda triggered by SNS message from CloudWatch alarm. Includes
deploy workflow and the executable code.

## ℹ️ Context

DPC is no longer allowed to use ChatBot to send messages to dpc-alerts
from AWS, so we are following a recommendation to have alerts trigger a
lambda that sends them to a slack webhook.

The infrastructure still builds S3 buckets, although this is not used
for actual deployment.

## 🧪 Validation

Workflow which tests code and terraform passes.
Deployed to dev manually and triggered manually and it worked.