PLT-1207: Remove github-actions-runner code and references (#264)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1207

## 🛠 Changes

Removes a bunch of code that we are no longer using since the move to
using Codebuild runners.

## ℹ️ Context

We no longer need this code since we shifted to using the AWS-provided
Codebuild runners. As part of the Greenfield migration cleanup, we're
getting rid of all references to the github-actions-runner
infrastructure.

## 🧪 Validation

We already don't use this, so there should be minimal changes outside of
the RDS terraform that references a legacy security group