PLT-1214 script to delete secrets from secrets manager (#266)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1214

## 🛠 Changes

Added a script to delete all AWS Secrets Manager secrets containing
ab2d, bcda, or dpc in the name.
Includes support for a --dry-run mode to safely preview deletions.


## ℹ️ Context

We are transitioning our secret management approach to standardized SOPs
using SSM Parameter Store instead of AWS Secrets Manager. As part of
this transition, we are cleaning up legacy secrets that are no longer
used. This script facilitates that cleanup in a safe and controlled way.

## 🧪 Validation

Tested locally using the --dry-run flag to confirm which secrets would
be deleted.

Co-authored-by: oluwolenpbc <boluwole@brians-MacBook-Pro-4.local>