Jd/dpc 4821 opt out use only cluster (#275)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4821

## 🛠 Changes

- Use aws_rds_cluster instead of aws_db_instance to find endpoint for
DB_HOST environment variable for opt out lambdas
- Remove ab2d from opt out lambdas as no longer use
- Workflows to update lambdas only apply to app:env combination that
have migrated to aurora

## ℹ️ Context

We are incrementally moving our databases to aurora clusters, which
require a slightly different terraform resource reference. As the lambda
infrastructure is stable, we are making the change in the existing
services and blocking deploy except to aurora environments.
App-specific github workflows to allow bcda to use at their own pace
when they are ready.

## 🧪 Validation

Manually ran in test environment, and testing the lambdas worked.

---------

Co-authored-by: Sean Fern <seanfern@navapbc.com>