PLT-1099: SOPS and SOPS Compatability (#270)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1099

## 🛠 Changes
Introduces a SOPS module that is _largely_ the same BFD's 01-config
module, as of v2.211.0 that's largely self-contained, acting as a
distribution method for the wrapper `sopsw` script that facilitates
adoption of the getsops.io `sops` binary.

Some changes to the related `standards` and `platform` modules could not
easily be decoupled from the primary changes toward the introduction of
the `sops` module itself. As a result, they are part of this change set.

As of this writing, the standards and documentation for managing SSM
Parameters with this module are still forthcoming. However, early
adoption of this module in AB2D has seen adoption of the loose standards
that BFD has been using for the past two years, namely:
* managed parameters are tagged with a set standards tags that include,
at a minimum, the team/app responsible for managing the value, the
environment to which the parameter is relevant, and a helpful link to
the infrastructure-as-code source code at issue in managing the specific
parameter resource
* all hierarchies in SSM Parameter Store include a 4-tuple prefix of
`/${app}/${env}/${service}/${sensitivity}` where:
* `app` is the application or team responsible for managing the
configuration
* `env` is the environment to which this value belongs and signifies the
inherited security posture/controls to be inherited
* `service` is generally the name of the service or terraservice to
which the configuration value applies
* `sensitivity` is either "sensitive" or "nonsensitive", and determined
whether or not the value should be treated as a secure string and
encrypted with the appropriate CMK
* value are encrypted with environment and app-specific KMS CMKs

## ℹ️ Context
Post greenfield migration, CDAP customers will be encouraged to adopt a
standard method for managing configuration values that are stored in AWS
SSM Parameter Store. This module should facilitate adoption of standards
for managing those values securely.

## 🧪 Validation
This module has been extensively validated in CDAP customer AB2D's
repository where a configuration-focused module and terraservice acts as
the primary entry point for environmental configuration values.

---------

Co-authored-by: Julia Reynolds <juliareynolds@navapbc.com>
Co-authored-by: Sean Fern <seanfern@navapbc.com>