PLT-1226: Extend alarm-to-slack service to cover all applications (#278)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1226

## 🛠 Changes

alarm-to-slack service was extended to cover all applications

## ℹ️ Context

the is the need for one instance of the alarm-to-slack function in each
account should be able to handle routing of alarms to Slack for all
applications and environments.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

see checks

---------

Co-authored-by: Maboh Christopher <chris.maboh@aquia.io>