PLT-1282: Update bucket module to use shared KMS keys (#295)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1282

## 🛠 Changes

bucket module was updated to use our shared kms keys

## ℹ️ Context

The bucket module currently creates KMS keys for encrypting and
decrypting tfstate buckets, the is the need to switch to the shared keys
created by the kms-keys service

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

See checks

---------

Co-authored-by: Maboh Christopher <chris.maboh@aquia.io>