PLT-1325: Add Prod env for opt out Lambdas for BCDA and DPC (#300)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1325

## 🛠 Changes

prod env was added to bcda and dpc opt-out lambdas

## ℹ️ Context

prod environment had been removed for previous work around aurora
migration, with the work now done the is need to add the prod
environment to workflow back

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation
See checks

---------

Co-authored-by: Maboh Christopher <chris.maboh@aquia.io>