[PLT-1108] Updating web module README sample usage commit hashes refs. (#319)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1108

## 🛠 Changes

This PR updates the web module README sample usage section with commit
hashes instead of branch references now that the branch has been merged.

## ℹ️ Context

The CDAP web module contains a sample usage snippet in the README that
contains three references to the branch in which changes were being
made. Since that branch has now been merged to main these references
should now be updated to the commit hash of that merge.

## 🧪 Validation

This is a README update that does not require validation.