PLT-1388: Add Sandbox Environment Back to Deploy Matrix for github-actions-role (#329)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1388

## 🛠 Changes

Adds sandbox back to the deploy matrix for `github-actions-role`.

## ℹ️ Context

We had a code freeze on the sandbox environment for the pentesting that
ended 10/10/2025. This PR aims to add the sandbox environment back to
the deploy matrix to reflect changes that were made to the permissions
of the `github-actions-role`, bringing it into parity with the other
environments.

## 🧪 Validation
This change has been validated in
https://github.com/CMSgov/cdap/pull/321