[PLT-1278] Adding SQL views and exports for AB2D insights dashboard. (#328)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1278

## 🛠 Changes

This PR consists of the eleven SQL view and export queries that map to
the eleven QuickSight datasets that comprise the AB2D DASG Insights
dashboard.

## ℹ️ Context

These queries have been run against the PROD account AB2D database to
create a set of views containing the data required by each dataset and
export it to an S3 bucket (also in the PROD account) that is accessible
by the QuickSight service in the DASG account.

## 🧪 Validation

These queries have already been run against the AB2D prod database - the
point of this PR is to preserve them for future use.