[PLT-1410] Adding pg_cron to Aurora module cluster parameter group. (#335)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1410

## 🛠 Changes

The Aurora DB module is being updated to enable pg_cron in the cluster
parameter group by default.

## ℹ️ Context

With AB2D, BCDA and DPC all now requiring pg_cron in at least the PROD
environment it makes sense to enable it by default in the Aurora module,
as is already the case with force_ssl.

## 🧪 Validation

A tofu plan against BCDA Sandbox with the updated module ref indicates
that the parameter will be added to the cluster parameter group:

```
# module.db.aws_rds_cluster_parameter_group.this will be updated in-place
  ~ resource "aws_rds_cluster_parameter_group" "this" {
        id          = "bcda-sandbox-cluster"
        name        = "bcda-sandbox-cluster"
        tags        = {}
       
      + parameter {
          + apply_method = "pending-reboot"
          + name         = "shared_preload_libraries"
          + value        = "pg_stat_statements,pg_cron"
        }
    }
```