[PLT-1227] Cost Anomaly alert to slack (#314)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1227

## 🛠 Changes

Added a cost anomaly cost monitor and sns topic subscription. This is
the first part of required work for this feature. A lambda to push
alerts to slack will be completed in PLT-1361

## ℹ️ Context

Users want to be alerted in Slack about significant cost fluctuations,
so that we can take actions on addressing potential cost anomalies.

## 🧪 Validation

See check here:  https://github.com/CMSgov/cdap/actions/runs/19306698285

---------

Co-authored-by: jdettmannnava <145699825+jdettmannnava@users.noreply.github.com>
Co-authored-by: Sean Fern <seanfern@navapbc.com>