PLT-1333: Remove backup_retention_period from aurora module (#344)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1333

## 🛠 Changes

Removes `backup_retention_period` and references to it.

## ℹ️ Context

We have an overlap with our snapshots and backup plans, meaning we can
remove the configuration from our aurora module.

## 🧪 Validation

Removing the variable and references to it may not disable the backup
retention entirely, as seen in this example of the output from pointing
the aurora module declaration in `dpc-ops` at this branch:
```
No changes. Your infrastructure matches the configuration.

OpenTofu has compared your real infrastructure against your configuration and found no differences, so no
changes are needed.
```